LayoutTests/canvas/philip/tests/2d.path.arc.selfintersect.1-expected.txt

< [index] >
2d.path.arc.selfintersect.1
arc() with lineWidth > 2*radius is drawn sensibly
References: 2d.path.arc.draw
Actual output:
Expected output:

Passed
